Add `DBShell::DatabaseClient`
A more generic interface to specific database adapter clients. This
class will act as a frontend to the others.

Calling its `.runshell()` will call the appropriate `.runshell()` for
the adapter provided in `connection_params`.

This way, we have a single entry point to start a database shell from
the Rake `dbshell` task.