Try to add support for shell aliases through Ruby's `Kernel.exec`
Work in progress.

Ruby's `system` and `exec` don't load shell aliases, so the executables
defined in `EXECUTABLE_NAME` will always be the ones used to launch
database shells.

The trouble is, users might have installed the command with a different
name (as in "mysql5"), or they might not have the command on their path
(as in my case when using Postgres.app).

Apparently you can use an interactive or login shell to get aliases
(using the `-l` or `-i` flags, which do slightly different things,
sourcing .profile and .bashrc separately). Here's what I was following:

http://stackoverflow.com/questions/12060863/python-subprocess-call-a-bash-alias/25099813#25099813

Unfortunately I haven't been able to get it to work. Dropping this for
now.